tenstorrent: bh_arc: initialize mrisc with priority 14
Initialize MRISC FW with priority 14. Note: this preserves relative order
since pcie_init() is run with priority 13.
